requirements.txt
Flask==2.3.3
flask-cors==4.0.0
google-cloud-vision==3.10.1
google-generativeai==0.3.2
python-dotenv==1.0.1